refactor: updates to EditionDialog
