+++ apps/videos-admin/src/app/[locale]/(dashboard)/videos/[videoId]/_VideoView/Editions/EditionDialog/index.ts
@@ -1,1 +1,1 @@
-export { EditionDialog, DialogAction } from './EditionDialog'
+export { EditionDialog } from './EditionDialog'
